refactor: 디렉토리 구조 절대경로로 변경
