+++ frontend/src/components/GoodsBlock/GoodsBlock.css
