+++ frontend/src/components/Slider/Slider.jsx
@@ -1,6 +1,6 @@
 import React, {useState, useEffect} from 'react';
-import {prevButtonIcon, nextButtonIcon} from '../constants/iconPath.js';
-import {SideDish} from './SideDish.jsx';
+import {prevButtonIcon, nextButtonIcon} from 'constants';
+import {SideDish} from 'components';
 
 function Slider({data}) {
   const [sliderState, setSliderState] = useState({clickedButton: '', list: ''});
