+++ frontend/src/components/index.js
@@ -1,0 +1,6 @@
+export * from './GoodsBlock/GoodsBlock';
+export * from './NavIcon/NavIcon';
+export * from './SideDish/SideDish';
+export * from './SideDishContentsList/SideDishContentsList';
+export * from './Slider/Slider';
+export * from './TabList/TabList';
