+++ frontend/src/constants/iconPath.js
