+++ frontend/src/containers/index.js
@@ -1,0 +1,3 @@
+export * from './Header/Header';
+export * from './SideDishContents/SideDishContents';
+export * from './SpecialPromotion/SpecialPromotion';
