+++ frontend/src/data/goods.js
