+++ frontend/src/data/index.js
@@ -1,0 +1,3 @@
+export * from './bestGoods';
+export * from './category';
+export * from './goods';
